pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0", 
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "moya-ai"
version = "0.1.4"
description = "A reference implementation of the Moya AI framework."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name="MontyCloud", email="kannan@montycloud.com" }
]
keywords = ["python", "moya"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "boto3>=1.36.9",
  "crewai>=0.100.1",
  "crewai-tools>=0.33.0",
  "requests>=2.32.3",
  "openai>=1.60.1",
  "fastapi>=0.115.7",
  "uvicorn>=0.34.0",
  "python-dotenv>=1.0.1",
  "langchain>=0.3.20",
  "anthropic>=0.49.0",
  "tokenizers>=0.21.1",
  "langchain-chroma>=0.2.2",
  "ollama>=0.4.7",
  "langchain_ollama>=0.2.3",
  "langchain_community>=0.3.19",
  "faiss-cpu>=1.10.0"
]
requires-python = ">=3.11" 

[tool.setuptools]
packages = [
    "moya",
    "moya.agents",
    "moya.classifiers",
    "moya.conversation",
    "moya.memory",
    "moya.orchestrators",
    "moya.registry",
    "moya.tools",
    "moya.utils",
    "moya.vectorstore",
    "moya.embeddings"
    ]
